Update README.md
26.	BLKN 492 Directed Research in Blockchain Studies 
BLKN 492 area of research.
Academic Adviser

The purpose of this course is to aid in the creation of your Capstone Projects. This is an opportunity for students to engage in advanced research in a professional field of specialization. The capstone projects will be inquiry- and practice-based and will be based on student- and program-specific areas of interest. The student completes a final original research project for the nanodegree in this directed research course. In addition to the advice provided by your course professors, your academic adviser from your particular field will provide additional help and comments throughout this phase of the study. Throughout the cohort, students will submit preliminary versions of the project's components to the academic adviser. The final output includes an in-depth written report of the inquiry and a formal presentation on the subject to the capstone defense committee. Students are expected to be able to respond to inquiries regarding their study and engage in professional dialogue on the issue.